keyring: split structs to its own file
